Move CreatePipeEx to WinEx
